requirements.txt
aiogram==3.24.0          # или новее
asyncpg
python-dotenv